requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.3
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyodbc==5.3.0
python-dotenv==1.2.1
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
zipp==3.23.0
